Make tests wait for the right condition
