+++ gaphor/ui/tests/test_modelchanged.py
@@ -2,11 +2,12 @@ from gi.repository import GLib
 
 from gaphor.core.modeling import ModelReady
 from gaphor.ui.modelchanged import ModelChanged
+from gaphor.ui.tests.fixtures import iterate_until
 
 
-def iteration():
+def iteration(sentinel):
     ctx = GLib.main_context_default()
-    while ctx.pending():
+    while ctx.pending() and not sentinel:
         ctx.iteration(False)
 
 
@@ -27,7 +28,7 @@ def test_monitor_file_changed(event_manager, tmp_path):
     event_manager.handle(ModelReady(None, filename=new_file))
 
     new_file.write_text("b", encoding="utf-8")
-    iteration()
+    iterate_until(widget.get_revealed)
 
     assert widget.get_revealed()
 
@@ -41,6 +42,6 @@ def test_monitor_file_not_changed(event_manager, tmp_path):
     event_manager.handle(ModelReady(None, filename=new_file))
 
     new_file.write_text("a", encoding="utf-8")
-    iteration()
+    iterate_until(condition=lambda: False, timeout=2)
 
     assert not widget.get_revealed()
